Refatorando Entidades e Relacionamentos

+++ backend/movimentos-manuais-api/src/main/java/com/bnpparibas/movimentosmanuais/domain/repository/MovimentoManualRepository.java
@@ -7,6 +7,6 @@ import com.bnpparibas.movimentosmanuais.domain.model.MovimentoManual;
 
 
 @Repository
-public interface MovimentoManualRepository extends JpaRepository<MovimentoManual, Integer>{
+public interface MovimentoManualRepository extends JpaRepository<MovimentoManual, Long>{
 
 }
