+++ backend/movimentos-manuais-api/src/main/java/com/bnpparibas/movimentosmanuais/domain/service/MovimentoManualService.java
@@ -10,7 +10,7 @@ public interface MovimentoManualService {
 
 	
 	public MovimentoManual salvar(MovimentoManual movimentoManual);
-	public void excluir(Integer id);
-	public Optional<MovimentoManual> findById(Integer id);
-	public boolean existsById(@Valid Integer id);
+	public void excluir(Long id);
+	public Optional<MovimentoManual> findById(Long id);
+	public boolean existsById(@Valid Long id);
 }
